corba: allow to use both the TAO and OmniORB CORBA implementations
The TAO implementation is really taking too much resources for my need.
On the other hand, people I know are very happy about OmniORB for a
robot control application, so I made the choice of trying OmniORB.

This patch allows to specify the desired Corba implementation (TAO or
OmniORB) at compile-time by specifying the CORBA_IMPLEMENTATION CMake
variable. The default choice is still TAO of course.

A specific header, the rtt-corba-config.h is now installed to tell which
implementation is in use. The CORBA_IS_TAO and CORBA_IS_OMNIORB
variables are defined accordingly.